Adds the ability to load scripts to tracking links, sensor boosters, sensor damps, tracking computers and tracking disruptors
If you wish to use this feature it REQUIRES that you add the scripts to your saved fitting and that you use the fitting manager.
ARM does NOT load scripts into your cargo they are put in your cargo by the fitting screen and your saved fitting!

this can likely be extended to support reloading the ancillary shield booster.... at a later date