This adds support for reloading and using Ancillary Shield Boosters
this reloads cap injectors but does not yet use them in combat!

THIS NEEDS TESTING, DO NOT AFK with an ancillary shield booster until you know it works! you have been warned.
YOU HAVE TO ADD YOUR CAP CHARGES TO YOUR SAVED FITS! YOU HAVE TO HAVE FITTING SUPPORT ON! Arm does NOT and will not put the cap booster charges in your cargo the fitting manager of eve MUST do it for you! make sure you test multiple missions in a row while watching it! This is a new type of module and needs extra scrutiny.... please file issues on github if you see any