Move some logging to only show when DebugCleanup = true
Change the dialog
"
If you decline of fail a mission from an agent he/she might become displeased and lower your standing towards him/her. You can decline a mission every four hours without penalty
"
from a sayyes to a close dialog, as it only has a close button.
If their is another dialog with a similar set of text with a yes button I have not seen it.