Adds new dependency, that I wish I could avoid... HTMLAgility allows us to read the HTML from the missions more easily.
Specifically it allows us to get the DungeonID from the HTML, then we can log it and eventually create a database of missions with unique identifiers (not just mission name, but the ACTUAL mission identifier, the DungeonID), this will eventually fix some issues detecting the 'right' faction, etc... if we know the DungeonID we know what faction its against, we know the NPC layout, etc, etc. if we don't then we are just guessing what faction if we cant detect the faction from the icon in the html, which happens a decent percentage of the time.

the new dependency will automatically be copied to the output directory when compiling, and automatically be copied to the .net programs directory when using #transfertolive#.bat, if copying manually make sure you grab the new DLL...