When referencing Stargates, ClosestStargate, bigObjects, etc make sure that we re-query DE and do not try to use a cached, possibly incorrect result
this is specifically because if we do not do this for "Stargates" we end up thinking the "closest" stargate is the one that we were next to when we entered system, even though we've already warped across the system and are sitting at a totally diff (and very close) stargate. This resolves that bug.