Changes Inventory functions along with Arm and Unloadloot to not use the Ammo window or Loot Window.
We now arm and Unloadloot without having to open the corresponding hangar window. We have been ignoring the window itself for a long time but unnecessarily checked for its existence.
The Cargo and ItemHangar still works the way it always has and does open an inventory window. This guarantees we always have an inventory window open.