Change the Dedicated Salvager to not depend on the UnloadLootAtStation setting. This is a dedicated salvager, lets assume they want all the loot and that back and fourth to the station is ok.
This also adds a timeout to a salvaging session (pocket) of 20 min. Any pocket that takes longer than that will trigger the dedicated salvager to go back to base. THis should keep the salvager from EVER getting hung for too long (though it should not get hung at all, if you notice any issues with this kicking in I'll want to see the logs)