rearrange the debug* settings in settings.cs to be sorted alphabetically.
this also adds debugDefense and moves two logging messages to only be shown when debugDefense is true.

This means that we will no longer have to see the log messages when we jump from system to system and our hardeners have to turn on one at a time at every session change.