Adds a new setting: "MoveCommonMissionCompletionItemsToItemsHangar" if set to true (the default is true) common mission completion items will be moved to your local hangar
this allows them to automatically be used to complete missions and should be left at true unless you have a VERY good reason to change it.
This setting is useful if you use the corp hangar for your AmmoHangar and LootHangar, if the corp hangar were configured for both of those there was previously no way to get these items back into your local items hangar by default.