Move TravelToAgentStation to travel.cs, create TravelHome function
remove local TravelToAgentStation functions and use the new shared function... if you use Travel.TravelHome() is will check to see if you are in CombatMissionsBehavior and if so there will be no change in behavior.
If you are not then it will use your HomeBookmarkName (eventually, but not yet)... at the moment either way it uses the same function, they are just named differently. In the next iteration the HomeBookmarkName will actually be used to find "home" for 'other' behaviors, including the DedicatedSalvager