Fix for: https://github.com/ISeeDEDPpl/Questor/issues/255
only align to the star in a non-speed tanked hybrid or projectile boat that moves less than 300m/s... NavigateIntoRange only calls OrbitgateorTarget if you are speed tanking OR if you are not speed tanking and are shooting a frigate. This should make killing frigates easier as the transversal will not be as high as it is when orbiting the target.