ValueDump and the ValueDump action within QuestorManager now use the same common code in Actions\Market.cs
Both have been tested and are working. If their are any corner case bugs please open an issue on github with logs. Keep in mind that both versions of valuedump will use a character settings xml now and you should have one for every toon that will use valuedump (though not strictly needed because it will use sane defaults without a config)...
If you are seeing any issues that need attention add ,<ebugValuedump>true</debugValuedump> to your characters config xml and link the logs to an issue on github.