Add ISXStealth and a script that gets called by questor BEFORE using directeve. If you use #compile#.bat and #transfertolive#.bat everything will be put in place for you.
If you want to do it by hand: copy isxstealth.dll to the innerspace\extensions directory and the StartingQuestorSafetyScript.iss to your innerspace\scripts directory.

The script is currently required. If the script is not found in the scripts directory questor will hang with an error message on the console before logging in telling you that the script is missing.
READ [ Questor now saves you from shooting yourself in the foot ]
If you already load isxstealth you do NOT need to make any changes to your scripts, this StartingQuestorSafetyScript cgecks to make sure isxstealth is not loaded and only loads it if needed.