When using the drop action keep track of how many times we have tried to open the cargo of the target and error after 5 tries.
If it does go into error (looping) it will spam the console log and keep running defense (and combat) but will more or less be stopped. this should only happen if we dont know how to detect the container is open or the container doesnt allow you to open it at this time

Is their a better solution to this?