Approaching/Orbiting is working better now.
For some reason Priority targets are not being set. Panic is trying but they are not making it into the list. The use of this without a fix is not advised unless you put all of the EWAR in kill commands for now. Committing this due to the major other fixes in getbest which are making things hela smoother combined with deds earlier commits.