lets see how much we can break with one commit, k?
refactored GetBestTarget and GetBestDroneTarget to use FindPrimaryWeaponPriorityTarget() and FindDronePriorityTarget()
fixed some bugs in getBesttarget where it was taking targets that were targeting (not yet .istarget)! bad bad bad
added a timer to the GetBestTarget and only reassign PreferredPrimaryWeaponTarget if it is over 6 sec since our choice OR the PreferredPrimaryWeaponTarget is dead.
and TONS of other fixes / cleanup / simplifications...
Added debugAddDronePriorityTarget and debugAddPrimaryWeaponPriorityTarget settings if needed for debugging purposes