fixes Entitycache.Dock()
ModuleCache.ReloadAmmo() and ModuleCache.ChangeAmmo() to all be bool instead of void
The logging MIGHT be a bit much at the moment, we shall see.
If it is we can adjust some timestamps or put some of the debug logging behind setting(s) instead of defaulting to on.
Improved logging == good