We now have MUCH more info on ammo attributes and can determine when missiles will hit...
FYI this defaults to off. To test enable: svoidShootingTargetsWithMissilesIfweKNowTheyAreAboutToBeHitWithAPreviousVolley in your settings XML
If this works well I'll start working on calculating # of volleys to kill an NPC and we can improve this a bit.
There are probably some rigs and ship bonuses that effect missile speed.... if you are using those I think this calculation will be off. (by more than 1/2 a sec though, I do not know)