OK, this is a work in progress and does not currently work!
If you use AdaptEVE to load QuestorDLL.dll you are told the signature doesnt match what its expecting. I pretty much gave up on that and wanted to start simple so in the QuestorDLL directory you have the QuestorLoader.csproj that spits out QuestorLoader.dll.

Note: As soon as I add an instance of the questor class I have to add a reference for DirectEVE (makes sense) which then confuses things (double references to the Easyhook hooks as DE seems to have em too)